requirements.txt
torch>=1.9.0
torchvision>=0.10.0
hydra-core>=1.1.0
omegaconf>=2.1.0
wandb>=0.12.0
tqdm>=4.62.0
datasets>=1.16.0
transformers>=4.20.0
numpy>=1.21.0
pandas>=1.3.0 